+++ egui-winit/src/window_settings.rs
@@ -1,7 +1,6 @@
-use egui_winit::winit;
-
+/// Can be used to store window settings (position and size).
 #[derive(Clone, Copy, Debug)]
-#[cfg_attr(feature = "persistence", derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct WindowSettings {
     /// outer position of window in physical pixels
     pos: Option<egui::Pos2>,
@@ -10,14 +9,8 @@ pub struct WindowSettings {
 }
 
 impl WindowSettings {
-    #[cfg(feature = "persistence")]
-    pub fn from_ron_file(settings_ron_path: impl AsRef<std::path::Path>) -> Option<WindowSettings> {
-        crate::persistence::read_ron(settings_ron_path)
-    }
-
-    pub fn from_window(window: &glutin::window::Window) -> Self {
-        let scale_factor = window.scale_factor();
-        let inner_size_points = window.inner_size().to_logical::<f32>(scale_factor);
+    pub fn from_display(window: &winit::window::Window) -> Self {
+        let inner_size_points = window.inner_size().to_logical::<f32>(window.scale_factor());
 
         Self {
             pos: window