+++ egui_demo_lib/src/apps/demo/dancing_strings.rs
@@ -1,8 +1,8 @@
 use egui::{containers::*, *};
 
 #[derive(Default)]
-#[cfg_attr(feature = "persistence", derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "persistence", serde(default))]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "serde", serde(default))]
 pub struct DancingStrings {}
 
 impl super::Demo for DancingStrings {
@@ -11,7 +11,7 @@ impl super::Demo for DancingStrings {
     }
 
     fn show(&mut self, ctx: &CtxRef, open: &mut bool) {
-        use super::View;
+        use super::View as _;
         Window::new(self.name())
             .open(open)
             .default_size(vec2(512.0, 256.0))