+++ egui_demo_lib/src/apps/demo/window_options.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, PartialEq)]
-#[cfg_attr(feature = "persistence", derive(serde::Deserialize, serde::Serialize))]
+#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct WindowOptions {
     title: String,
     title_bar: bool,
@@ -55,7 +55,7 @@ impl super::Demo for WindowOptions {
             ctx.request_repaint();
         }
 
-        use super::View;
+        use super::View as _;
         let mut window = egui::Window::new(title)
             .id(egui::Id::new("demo_window_options")) // required since we change the title
             .resizable(resizable)